+++ molecularnodes/entities/__init__.py
@@ -1,14 +1,19 @@
-from . import molecule, trajectory
+from .molecule import CLASSES as MOL_CLASSES
+from .trajectory import CLASSES as TRAJ_CLASSES
 from .density import MN_OT_Import_Map
 from .trajectory.dna import MN_OT_Import_OxDNA_Trajectory
+from .ensemble.ui import MN_OT_Import_Cell_Pack, MN_OT_Import_Star_File
+
 from .ensemble.cellpack import CellPack
 from .ensemble.star import StarFile
-from .ensemble.ui import MN_OT_Import_Cell_Pack, MN_OT_Import_Star_File
 from .molecule.pdb import PDB
 from .molecule.pdbx import BCIF, CIF
 from .molecule.sdf import SDF
 from .molecule.ui import fetch, load_local
-from .trajectory.trajectory import Trajectory
+from .trajectory import Trajectory
+from .molecule import Molecule
+from .ensemble import Ensemble
+
 
 CLASSES = (
     [
@@ -17,6 +22,6 @@ CLASSES = (
         MN_OT_Import_OxDNA_Trajectory,
         MN_OT_Import_Star_File,
     ]
-    + trajectory.CLASSES
-    + molecule.CLASSES
+    + TRAJ_CLASSES
+    + MOL_CLASSES
 )
