+++ molecularnodes/entities/ensemble/__init__.py
@@ -1,1 +1,2 @@
 from .ui import load_starfile, load_cellpack
+from .ensemble import Ensemble
