+++ molecularnodes/entities/ensemble/ensemble.py
@@ -18,11 +18,9 @@ class Ensemble(MolecularEntity, metaclass=ABCMeta):
 
         """
         super().__init__()
-        self.type: str = "ensemble"
         self.file_path: Path = bl.path_resolve(file_path)
         self.instances: bpy.types.Collection = None
         self.frames: bpy.types.Collection = None
-        bpy.context.scene.MNSession.ensembles[self.uuid] = self
 
     @classmethod
     def create_object(
